Moved to util.stream package.
