+++ core/src/main/java/com/google/mu/util/stream/package.html
@@ -1,0 +1,16 @@
+<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 3.2 Final//EN">
+<html>
+<head>
+Stream Utilities
+</head>
+<body bgcolor="white">
+
+Stream utilities.
+<br>
+
+<h2>Package Specification</h2>
+
+The code is developed and tested under Java 8.0. 
+
+</body>
+</html>
