+++ core/src/test/java/com/google/mu/util/stream/MoreStreamsParameterizedTest.java
@@ -1,4 +1,4 @@
-package com.google.mu.util;
+package com.google.mu.util.stream;
 
 import static com.google.common.truth.Truth.assertThat;
 import static java.util.Arrays.asList;
@@ -14,6 +14,8 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
+
+import com.google.mu.util.stream.MoreStreams;
 
 @RunWith(Parameterized.class)
 public class MoreStreamsParameterizedTest {
