+++ core/src/test/java/com/google/mu/util/stream/MoreStreamsTest.java
@@ -1,4 +1,4 @@
-package com.google.mu.util;
+package com.google.mu.util.stream;
 
 import static com.google.common.truth.Truth.assertThat;
 import static java.util.Arrays.asList;
@@ -17,6 +17,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
 import com.google.common.testing.ClassSanityTester;
+import com.google.mu.util.stream.MoreStreams;
 
 @RunWith(JUnit4.class)
 public class MoreStreamsTest {
